Using ExampleTemplate for examples/code accessibility.
