+++ packages/terra-data-grid/src/DataGrid.module.scss
@@ -27,6 +27,7 @@
       display: flex;
       flex: 0 0 auto;
       height: 100%;
+      z-index: 1;
     }
 
     .header-overflow-container {
