+++ packages/terra-data-grid/src/subcomponents/Cell.module.scss
@@ -20,7 +20,6 @@
 
     &.selectable:hover {
       background-color: var(--terra-clinical-data-grid-cell-hover-background-color, #e2f2fc);
-      box-shadow: inset 0 0 5px 0 #a6d4fc;
       cursor: pointer;
     }
