+++ packages/terra-data-grid/src/subcomponents/ResizeHandle.module.scss
@@ -13,7 +13,8 @@
 
     &.dragging {
       background-color: var(--terra-clinical-data-grid-resize-handle-active-background-color, darkslategrey);
-      z-index: 200;
+      // transform: translateZ(0);
+      // z-index: 200;
     }
   }
 }
