+++ packages/terra-data-grid/src/subcomponents/Row.module.scss
@@ -16,3 +16,7 @@
   background-color: var(--terra-clinical-data-grid-row-selected-background-color, #cbe7fa);
   box-shadow: inset 0 0 5px 0 #0092e0;
 }
+
+.row.selected.hover {
+  background-color: var(--terra-clinical-data-grid-row-hover-background-color, #b3ddf7);
+}
