+++ packages/terra-data-grid/src/terra-dev-site/doc/data-grid/Examples/DatagridWithPaging.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 
-import Button from 'terra-button';
 // eslint-disable-next-line import/no-extraneous-dependencies, import/no-unresolved, import/extensions
 import DataGrid from 'terra-data-grid';
+import LoadingOverlay from 'terra-overlay/lib/LoadingOverlay';
 
 import ContentCellLayout from './ContentCellLayout';
 
@@ -45,24 +45,9 @@ const overflowColumns = [
     width: 200,
     text: 'Column 6',
   },
-  {
-    id: 'Column-7',
-    width: 200,
-    text: 'Column 7',
-  },
-  {
-    id: 'Column-8',
-    width: 200,
-    text: 'Column 8',
-  },
-  {
-    id: 'Column-9',
-    width: 200,
-    text: 'Column 9',
-  },
 ];
 
-class SubsectionDataGrid extends React.Component {
+class DatagridWithPaging extends React.Component {
   static buildRows(sectionId, num) {
     const rows = (new Array(num)).fill().map((rowVal, rowIndex) => ({
       id: `${sectionId}-Row${rowIndex}`,
@@ -75,52 +60,58 @@ class SubsectionDataGrid extends React.Component {
     return rows;
   }
 
+  static buildSections(sectionCount) {
+    const sections = [];
+    for (let i = 0, length = sectionCount; i < length; i += 1) {
+      const sectionId = `section_${i}`;
+      sections.push({
+        id: sectionId,
+        text: `Section ${i}`,
+        rows: DatagridWithPaging.buildRows(sectionId, 3),
+      });
+    }
+
+    return sections;
+  }
+
+
   constructor(props) {
     super(props);
 
-    this.buildSection = this.buildSection.bind(this);
-
     this.state = {
-      collapsedSectionId: undefined,
+      sectionCount: 1,
+      isLoading: false,
+      sections: DatagridWithPaging.buildSections(1),
     };
   }
 
-  buildSection(sectionId, sectionName, numberOfRows) {
-    return {
-      id: sectionId,
-      text: sectionName,
-      endAccessory: (sectionId === 'section_1') ? <span><Button text="Button 1" data-accessible-data-grid-content /><Button text="Button 2" data-accessible-data-grid-content /></span> : null,
-      isCollapsible: sectionId === 'section_0',
-      isCollapsed: this.state.collapsedSectionId === sectionId,
-      rows: SubsectionDataGrid.buildRows(sectionId, numberOfRows),
-    };
+  componentWillUnmount() {
+    clearTimeout(this.pagingTimeout);
   }
 
   render() {
     return (
-      <div style={{ height: '100%', padding: '15px' }}>
+      <div style={{ height: '800px', position: 'relative' }}>
         <DataGrid
           pinnedColumns={pinnedColumns}
           overflowColumns={overflowColumns}
-          sections={[
-            this.buildSection('section_0', 'Section 0', 15),
-            this.buildSection('section_1', 'Section 1', 15),
-            this.buildSection('section_2', 'Section 2', 15),
-          ]}
+          sections={this.state.sections}
           rowHeight="2.5rem"
           headerHeight="3rem"
-          onRequestSectionCollapse={(sectionId) => {
-            if (this.state.collapsedSectionId === sectionId) {
-              this.setState({ collapsedSectionId: undefined });
-            } else {
-              this.setState({ collapsedSectionId: sectionId });
-            }
-          }}
           fill
+          onRequestContent={this.state.sectionCount < 10 ? (() => {
+            this.setState({ isLoading: true }, () => {
+              clearTimeout(this.pagingTimeout);
+              this.pagingTimeout = setTimeout(() => {
+                this.setState({ sectionCount: this.state.sectionCount + 1, isLoading: false, sections: DatagridWithPaging.buildSections(this.state.sectionCount + 1) });
+              }, 2000);
+            });
+          }) : undefined}
         />
+        {<LoadingOverlay isOpen={this.state.isLoading} isRelativeToContainer isAnimated style={{ top: 0 }} />}
       </div>
     );
   }
 }
 
-export default SubsectionDataGrid;
+export default DatagridWithPaging;
